move reg form to components
