+++ apps/consent-ui/src/app/[lang]/register/index.tsx
@@ -20,7 +20,7 @@
 import Link from 'next/link';
 import { getTranslation, ValidLanguage } from 'src/i18n';
 
-import RegistrationForm from './RegistrationForm';
+import RegistrationForm from 'src/components/RegistrationForm';
 
 const ParticipantRegistration = async ({ currentLang }: { currentLang: ValidLanguage }) => {
 	const translate = await getTranslation(currentLang);
